Downgraded to jQuery 1.8.
Our version of Bootstrap relies on $.browser, which is not available
in jQuery 1.9. We get jQuery from Google Hosted Libraries, and we were
just specifying "1.x", so when Google updated its hosted libraries,
Bootstrap broke.

There's probably a new version of Bootstrap that works with jQuery 1.9,
but for now I'm just specifying jQuery 1.8 so the site will work again.